requirements.txt
streamlit
fpdf
pypdf
faiss-cpu
langchain
